Compile files before publishing
Compiling files before publishing reduces extra config from user so that they won't need to configure Jest, Webpack etc. to compile a library manually. It's also useful when people are using it on the web with SSR because setting this up with Node can be tricky.

Metro will still use the source files since we point the `react-native` key to `src`, so it'll work the same as before.
