+++ .gitignore
@@ -49,3 +49,5 @@ android/keystores/debug.keystore
 
 # CocoaPods
 ios/Pods/
+# generated by bob
+lib/
