+++ package.json
@@ -8,12 +8,17 @@
     "test:unit": "jest",
     "format": "prettier --write --list-different './src/**/*.js'",
     "lint": "eslint './src/**/*.js'",
-    "release": "npm login && release-it"
+    "release": "npm login && release-it",
+    "prepare": "bob build"
   },
-  "main": "src/Animated.js",
+  "main": "lib/commonjs/Animated",
+  "module": "lib/module/Animated",
+  "react-native": "src/Animated",
+  "source": "src/Animated",
   "types": "react-native-reanimated.d.ts",
   "files": [
     "src/",
+    "lib/",
     "android/src/main/AndroidManifest.xml",
     "android/src/main/java/",
     "android/build.gradle",
@@ -50,6 +55,7 @@
     "@babel/plugin-proposal-class-properties": "^7.7.4",
     "@babel/preset-env": "^7.7.6",
     "@babel/preset-typescript": "^7.7.4",
+    "@react-native-community/bob": "^0.14.3",
     "@react-native-community/eslint-config": "^0.0.5",
     "@types/react": "^16.9.0",
     "@types/react-native": "0.60.19",
@@ -88,5 +94,13 @@
         "# check if version corresponds to changes in native files  \n git diff --name-only ${latestVersion} HEAD | egrep \"(android/.*)|(ios/.*)\" -q && egrep '\\.0$' -q <<< ${version}"
       ]
     }
+  },
+  "@react-native-community/bob": {
+    "source": "src",
+    "output": "lib",
+    "targets": [
+      "commonjs",
+      "module"
+    ]
   }
 }
