+++ src/ReanimatedModule.native.js
@@ -1,0 +1,5 @@
+import { NativeModules } from 'react-native';
+
+const { ReanimatedModule } = NativeModules;
+
+export default ReanimatedModule;
